refactor: Extract i18n types to an external file
