+++ app/[lang]/layout.tsx
@@ -3,9 +3,10 @@ import { cookies } from 'next/headers'
 import { getCookie } from 'cookies-next/server'
 import { WebSite, WithContext } from 'schema-dts'
 import { Analytics } from '@vercel/analytics/react'
+import type { Locale } from '@/types'
 import { PageProps, Theme } from '@/types'
 import { COOKIES, metadataBase } from '@/constants'
-import { Locale, i18n } from '@/i18n.config'
+import { i18n } from '@/i18n.config'
 import { getDictionary } from '@/scripts'
 import { getAbsoluteURL, getMetaImage } from '@/lib'
 import { montserrat, lora, firaCode } from '@/theme/fonts'
