+++ app/[lang]/not-found.tsx
@@ -1,6 +1,6 @@
 import { cookies } from 'next/headers'
 import { getCookie } from 'cookies-next/server'
-import { Locale } from '@/i18n.config'
+import type { Locale } from '@/types'
 import { COOKIES, LINKS } from '@/constants'
 import { getDictionary } from '@/scripts'
 import { localizePath } from '@/lib'
