+++ design-system/components/post/post.types.tsx
@@ -1,6 +1,5 @@
-import { Locale } from '@/i18n.config'
+import type { Locale, Post } from '@/types'
 import { Dictionary } from '@/scripts'
-import type { Post } from '@/types'
 
 export type PostProps = Pick<
   Post,
