+++ design-system/components/project/project.types.tsx
@@ -1,6 +1,5 @@
 import { Dictionary } from '@/scripts'
-import { Locale } from '@/i18n.config'
-import type { Project } from '@/types'
+import type { Locale, Project } from '@/types'
 
 export type ProjectProps = Pick<
   Project,
