+++ i18n.config.ts
@@ -1,12 +1,4 @@
-export interface I18nConfig {
-  defaultLocale: 'en'
-  locales: ['en', 'pl']
-  regions: ['US', 'PL']
-  region: Record<Locale, Region>
-}
-
-export type Locale = I18nConfig['locales'][number]
-export type Region = I18nConfig['regions'][number]
+import type { I18nConfig } from '@/types'
 
 export const i18n: I18nConfig = {
   defaultLocale: 'en',
