+++ middleware.ts
@@ -2,8 +2,9 @@ import { NextResponse } from 'next/server'
 import type { NextRequest } from 'next/server'
 import { match } from '@formatjs/intl-localematcher'
 import Negotiator from 'negotiator'
+import type { I18nConfig } from '@/types'
 import { COOKIES } from './constants'
-import { i18n, I18nConfig } from './i18n.config'
+import { i18n } from './i18n.config'
 
 function getLocale(request: NextRequest, i18nConfig: I18nConfig): string {
   const { locales, defaultLocale } = i18nConfig
